End implementing first version of bluerobotics ping1D Small correction of loop_dt (long instead of int)

+++ src/seabot2_driver/gpsd_client/src/gpsd_node.cpp
@@ -34,7 +34,7 @@ GpsdNode::~GpsdNode() {
 }
 
 void GpsdNode::init_parameters() {
-    this->declare_parameter<int>("loop_dt", loop_dt_.count());
+    this->declare_parameter<long>("loop_dt", loop_dt_.count());
     loop_dt_ = std::chrono::milliseconds(this->get_parameter_or("dt", loop_dt_.count()));
 
     this->declare_parameter<string>("frame_id", frame_id_);
